backend/src/rhythm/saju.py
"""
사주명리 계산 모듈

⚠️ 이 모듈은 기존 사주명리 계산 로직을 통합하기 위한 인터페이스입니다.

통합 방법:
1. 기존 사주명리 계산 코드를 이 파일에 추가
2. calculate_saju() 함수가 기존 코드를 호출하도록 수정
3. 출력 형식을 Dict[str, Any]로 통일

내부 전문 용어 사용 가능:
- 천간(天干), 지지(地支), 오행(五行), 십성(十星)
- 대운(大運), 세운(歲運), 월운(月運), 일운(日運)
- 천을귀인, 역마, 공망 등

⚠️ 이 계산 결과는 사용자에게 직접 노출하지 않습니다!
Content Assembly Engine에서 일반 언어로 변환됩니다.
"""
from datetime import date, time, datetime
from typing import Dict, Any, Optional
from .models import BirthInfo, RhythmSignal


def calculate_saju(birth_info: BirthInfo, target_date: date) -> Dict[str, Any]:
    """
    사주명리 계산 (기존 로직 통합 지점)

    Args:
        birth_info: 출생 정보
        target_date: 분석 대상 날짜

    Returns:
        사주명리 계산 결과 (내부 전문 용어 사용)

    TODO: 기존 사주명리 계산 라이브러리/코드를 여기에 통합하세요
    """
    # ============================================================================
    # [통합 지점] 기존 사주명리 계산 코드를 여기에 추가하세요
    # ============================================================================

    # 예시 구조 (실제 계산 로직으로 대체 필요)
    saju_result = {
        "사주": {
            "년주": {
                "천간": _get_heavenly_stem(birth_info.birth_date.year),
                "지지": _get_earthly_branch(birth_info.birth_date.year)
            },
            "월주": {
                "천간": _get_heavenly_stem(birth_info.birth_date.month),
                "지지": _get_earthly_branch(birth_info.birth_date.month)
            },
            "일주": {
                "천간": _get_heavenly_stem(birth_info.birth_date.day),
                "지지": _get_earthly_branch(birth_info.birth_date.day)
            },
            "시주": {
                "천간": _get_heavenly_stem(birth_info.birth_time.hour),
                "지지": _get_earthly_branch(birth_info.birth_time.hour)
            }
        },
        "오행": {
            "목": 0,
            "화": 0,
            "토": 0,
            "금": 0,
            "수": 0
        },
        "십성": {
            "비견": False,
            "겁재": False,
            "식신": False,
            "상관": False,
            "편재": False,
            "정재": False,
            "편관": False,
            "정관": False,
            "편인": False,
            "정인": False
        },
        "특수신살": {
            "천을귀인": [],
            "역마": False,
            "공망": False,
            "도화": False
        },
        # 대상 날짜의 일진 정보
        "target_date_info": {
            "일간": _get_heavenly_stem(target_date.day),
            "일지": _get_earthly_branch(target_date.day),
            "월령": _get_month_element(target_date.month)
        }
    }

    return saju_result


def _get_heavenly_stem(value: int) -> str:
    """
    천간 계산 (간이 버전)
    TODO: 정확한 천간 계산 로직으로 대체
    """
    stems = ["甲", "乙", "丙", "丁", "戊", "己", "庚", "辛", "壬", "癸"]
    return stems[value % 10]


def _get_earthly_branch(value: int) -> str:
    """
    지지 계산 (간이 버전)
    TODO: 정확한 지지 계산 로직으로 대체
    """
    branches = ["子", "丑", "寅", "卯", "辰", "巳", "午", "未", "申", "酉", "戌", "亥"]
    return branches[value % 12]


def _get_month_element(month: int) -> str:
    """
    월령 오행 계산 (간이 버전)
    TODO: 정확한 월령 계산 로직으로 대체
    """
    # 간단한 계절 구분
    if month in [3, 4, 5]:
        return "목(木)"
    elif month in [6, 7, 8]:
        return "화(火)"
    elif month in [9, 10, 11]:
        return "금(金)"
    else:  # 12, 1, 2
        return "수(水)"


def analyze_daily_fortune(
    birth_info: BirthInfo,
    target_date: date,
    saju_data: Dict[str, Any]
) -> Dict[str, Any]:
    """
    일간 운세 분석 (내부 해석)

    Args:
        birth_info: 출생 정보
        target_date: 분석 날짜
        saju_data: 사주명리 계산 결과

    Returns:
        일간 운세 해석 (내부 표현)
    """
    # ============================================================================
    # [통합 지점] 기존 일간 운세 해석 로직을 여기에 추가하세요
    # ============================================================================

    # 예시 구조
    fortune_analysis = {
        "에너지_수준": 3,  # 1-5
        "집중력": 4,
        "사회운": 3,
        "결정력": 4,
        "유리한_시간": ["오전 9-11시 (巳時)", "오후 2-4시 (申時)"],
        "주의_시간": ["오후 5-7시 (酉時)"],
        "유리한_방향": ["북동(艮)", "남서(坤)"],
        "주요_흐름": "안정과 정리의 날",
        "기회_요소": ["관계 강화", "학습 집중"],
        "도전_요소": ["충동 조절 필요"]
    }

    return fortune_analysis


def get_favorable_times(saju_data: Dict[str, Any], target_date: date) -> list[str]:
    """
    유리한 시간대 계산

    TODO: 시진(時辰) 기반 정확한 계산 로직으로 대체
    """
    # 간이 계산
    day_stem = saju_data.get("target_date_info", {}).get("일간", "甲")

    # 예시: 천을귀인 시간 계산 (실제 로직으로 대체 필요)
    return ["오전 9-11시", "오후 2-4시"]


def get_favorable_directions(saju_data: Dict[str, Any], target_date: date) -> list[str]:
    """
    유리한 방향 계산

    TODO: 기문둔갑 또는 사주 기반 방위 계산 로직으로 대체
    """
    # 간이 계산
    return ["북동", "남서"]


# ============================================================================
# 통합 가이드
# ============================================================================
"""
기존 사주명리 코드 통합 방법:

1. 라이브러리/패키지가 있는 경우:
   - requirements.txt에 추가
   - 이 파일에서 import
   - calculate_saju() 함수에서 라이브러리 호출
   - 결과를 Dict 형식으로 변환

2. 직접 작성한 코드가 있는 경우:
   - 코드를 이 파일에 복사
   - 함수/클래스를 적절히 배치
   - calculate_saju()에서 호출

3. 외부 API를 사용하는 경우:
   - API 클라이언트 코드 추가
   - 환경변수에 API 키 설정
   - calculate_saju()에서 API 호출

예시:
    # 기존 라이브러리 사용
    from your_saju_library import SajuCalculator

    def calculate_saju(birth_info: BirthInfo, target_date: date):
        calculator = SajuCalculator()
        result = calculator.calculate(
            birth_date=birth_info.birth_date,
            birth_time=birth_info.birth_time,
            gender=birth_info.gender.value
        )
        return result.to_dict()
"""
