backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.0
postgrest==0.16.0
psycopg2-binary==2.9.9

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0

# PDF Generation
weasyprint==60.2
Pillow==10.2.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2024.1

# Korean Calendar / Saju (placeholder - integrate existing library)
# saju-library==x.x.x  # Replace with actual library

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
python-dotenv==1.0.1
